pyproject.toml
[tool.poetry]
name = "hmc-snuh"
version = "0.1.0"
description = ""
authors = ["baruda <maider030@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10,<3.11"
numpy = "^1.26.4"
kaggle = "^1.6.6"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
matplotlib = "^3.8.3"
scikit-learn = "^1.4.1.post1"
librosa = "^0.10.1"
albumentations = "^1.4.0"
hydra-core = "^1.3.2"
tsfm-public = {git = "https://github.com/IBM/tsfm.git"}
transformers = {git = "https://github.com/huggingface/transformers.git"}
datasets = "^2.17.1"
torchvision = "^0.17.1"
timm = "^0.9.16"
lightning = "^2.2.1"
lightning-bolts = "^0.7.0"
neuralforecast = {git = "https://github.com/maxpumperla/neuralforecast.git"}
mamba = "^0.11.3"
packaging = "^24.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"
pytest = "^8.0.1"
wandb = "^0.16.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
